@@ -2,35 +2,35 @@ class TagsController < ApplicationController
   before_filter :authenticate_user!
 
   # GET /tags
-  # GET /tags.xml
+  # GET /tags.json
   def index
     @tags = current_user.tags.all
 
     respond_to do |format|
       format.html # index.html.erb
-      format.xml  { render :xml => @tags }
+      format.json { render json: @tags }
     end
   end
 
   # GET /tags/1
-  # GET /tags/1.xml
+  # GET /tags/1.json
   def show
     @tag = current_user.tags.find(params[:id])
 
     respond_to do |format|
       format.html # show.html.erb
-      format.xml  { render :xml => @tag }
+      format.json { render json: @tag }
     end
   end
 
   # GET /tags/new
-  # GET /tags/new.xml
+  # GET /tags/new.json
   def new
     @tag = current_user.tags.new
 
     respond_to do |format|
       format.html # new.html.erb
-      format.xml  { render :xml => @tag }
+      format.json { render json: @tag }
     end
   end
 
@@ -40,47 +40,47 @@ class TagsController < ApplicationController
   end
 
   # POST /tags
-  # POST /tags.xml
+  # POST /tags.json
   def create
     @tag = current_user.tags.new(params[:tag])
     @tag.user_id = current_user.id
 
     respond_to do |format|
       if @tag.save
-        format.html { redirect_to(@tag, :notice => 'Tag was successfully created.') }
-        format.xml  { render :xml => @tag, :status => :created, :location => @tag }
+        format.html { redirect_to @tag, notice: 'Tag was successfully created.' }
+        format.json { render json: @tag, status: :created, location: @tag }
       else
-        format.html { render :action => "new" }
-        format.xml  { render :xml => @tag.errors, :status => :unprocessable_entity }
+        format.html { render action: "new" }
+        format.json { render json: @tag.errors, status: :unprocessable_entity }
       end
     end
   end
 
   # PUT /tags/1
-  # PUT /tags/1.xml
+  # PUT /tags/1.json
   def update
     @tag = current_user.tags.find(params[:id])
 
     respond_to do |format|
       if @tag.update_attributes(params[:tag])
-        format.html { redirect_to(@tag, :notice => 'Tag was successfully updated.') }
-        format.xml  { head :ok }
+        format.html { redirect_to @tag, notice: 'Tag was successfully updated.' }
+        format.json { head :ok }
       else
-        format.html { render :action => "edit" }
-        format.xml  { render :xml => @tag.errors, :status => :unprocessable_entity }
+        format.html { render action: "edit" }
+        format.json { render json: @tag.errors, status: :unprocessable_entity }
       end
     end
   end
 
   # DELETE /tags/1
-  # DELETE /tags/1.xml
+  # DELETE /tags/1.json
   def destroy
     @tag = current_user.tags.find(params[:id])
     @tag.destroy
 
     respond_to do |format|
-      format.html { redirect_to(tags_url) }
-      format.xml  { head :ok }
+      format.html { redirect_to tags_url }
+      format.json { head :ok }
     end
   end
 end
