@@ -58,7 +58,7 @@ Timefliesby::Application.routes.draw do
   # You can have the root of your site routed with "root"
   # just remember to delete public/index.html.
   # Devise authentication requires it!
-  root :to => "welcome#index"
+  root :to => 'welcome#index'
 
   # See how all your routes lay out with "rake routes"
 
