@@ -3,11 +3,10 @@ require 'spec_helper'
 describe WelcomeController do
 
   describe "Not Sign in as user" do
-
-    describe "GET index" do
-      it "Render template" do
-        get :index
-        response.should render_template("index")
+    describe "GET 'index'" do
+      it "should be successful" do
+        get 'index'
+        response.should be_success
       end
     end
 
