@@ -3,32 +3,32 @@ require "spec_helper"
 describe TagsController do
   describe "routing" do
 
-    it "recognizes and generates #index" do
-      { :get => "/tags" }.should route_to(:controller => "tags", :action => "index")
+    it "routes to #index" do
+      get("/tags").should route_to("tags#index")
     end
 
-    it "recognizes and generates #new" do
-      { :get => "/tags/new" }.should route_to(:controller => "tags", :action => "new")
+    it "routes to #new" do
+      get("/tags/new").should route_to("tags#new")
     end
 
-    it "recognizes and generates #show" do
-      { :get => "/tags/1" }.should route_to(:controller => "tags", :action => "show", :id => "1")
+    it "routes to #show" do
+      get("/tags/1").should route_to("tags#show", :id => "1")
     end
 
-    it "recognizes and generates #edit" do
-      { :get => "/tags/1/edit" }.should route_to(:controller => "tags", :action => "edit", :id => "1")
+    it "routes to #edit" do
+      get("/tags/1/edit").should route_to("tags#edit", :id => "1")
     end
 
-    it "recognizes and generates #create" do
-      { :post => "/tags" }.should route_to(:controller => "tags", :action => "create")
+    it "routes to #create" do
+      post("/tags").should route_to("tags#create")
     end
 
-    it "recognizes and generates #update" do
-      { :put => "/tags/1" }.should route_to(:controller => "tags", :action => "update", :id => "1")
+    it "routes to #update" do
+      put("/tags/1").should route_to("tags#update", :id => "1")
     end
 
-    it "recognizes and generates #destroy" do
-      { :delete => "/tags/1" }.should route_to(:controller => "tags", :action => "destroy", :id => "1")
+    it "routes to #destroy" do
+      delete("/tags/1").should route_to("tags#destroy", :id => "1")
     end
 
   end
